+++ R/format_detect.R
@@ -35,7 +35,8 @@ format_detect <- function(data, list.prices = NULL, force = FALSE) {
     "newton",
     "pooltool",
     "presearch",
-    "shakepay"))
+    "shakepay",
+    "uphold"))
   
   data_exchanges <- paste0("data_", exchanges)
   
@@ -48,30 +49,39 @@ format_detect <- function(data, list.prices = NULL, force = FALSE) {
   # Generate logical condition to identify right exchange
   condition <- names(which(data.names == exchanges.cols))
   
-  if (length(condition) == 0) {
+  if (all(condition == c("CDC_exchange_rewards", "CDC_wallet"))) {
+    if (any(unlist(lapply(c("Supercharger", "Interest", "APR", "Rebate"), grepl, data$Description)))) {
+      condition <- "CDC_exchange_rewards"
+    } else if (any(unlist(lapply(c("Validator", "Auto Withdraw"), grepl, data$Description)))) {
+      condition <- "CDC_wallet"
+    }
+  } else if (length(condition) == 0) {
     stop("Could not identify the correct exchange automatically. ",
          "Please use the appropriate function or 'format_generic()'.")
+  } else if (length(condition) > 1) {
+    stop("Matches multiple exchange names. Please report this bug so it can be fixed.")
   }
   
   # Apply right function
   formatted.data <- switch(
     condition,
+    shakepay = {format_shakepay(data)},
+    newton = {format_newton(data)},
+    pooltool = {format_pooltool(data)},
+    CDC = {format_CDC(data)},
+    celsius = {format_celsius(data)},
     adalite = {format_adalite(data, list.prices = list.prices, force = force)},
     binance = {format_binance(data, list.prices = list.prices, force = force)},
     binance_withdrawals = {format_binance_withdrawals(data, list.prices = list.prices, force = force)},
     blockfi = {format_blockfi(data, list.prices = list.prices, force = force)},
-    CDC = {format_CDC(data)},
     CDC_exchange_rewards = {format_CDC_exchange_rewards(data, list.prices = list.prices, force = force)},
     CDC_exchange_trades = {format_CDC_exchange_trades(data, list.prices = list.prices, force = force)},
     CDC_wallet = {format_CDC_wallet(data, list.prices = list.prices, force = force)},
-    celsius = {format_celsius(data)},
     coinsmart = {format_coinsmart(data, list.prices = list.prices, force = force)},
     exodus = {format_exodus(data, list.prices = list.prices, force = force)},
     gemini = {format_gemini(data, list.prices = list.prices, force = force)},
-    newton = {format_newton(data)},
-    pooltool = {format_pooltool(data)},
     presearch = {format_presearch(data, list.prices = list.prices, force = force)},
-    shakepay = {format_shakepay(data)},
+    uphold = {format_uphold(data, list.prices = list.prices, force = force)},
   )
   
   formatted.data